requirements.txt
# StyleFit Requirements
# ===================

# Core ML/AI libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Computer Vision
Pillow>=9.0.0
opencv-python>=4.8.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Similarity search
faiss-cpu>=1.7.4
# For GPU support (optional):
# faiss-gpu>=1.7.4

# Web interface
streamlit>=1.25.0
streamlit-option-menu>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress bars and utilities
tqdm>=4.65.0
pathlib2>=2.3.0

# Data validation
pydantic>=2.0.0

# Logging and monitoring
loguru>=0.7.0

# Optional: For advanced features
# jupyter>=1.0.0
# ipywidgets>=8.0.0
# tensorboard>=2.13.0
